--- src/main/java/Command.java
@@ -1,6 +1,0 @@
-public abstract class Command<T> {
-    public void execute() throws DukeException {}
-    public boolean isExit(){
-        return true;
-    }
-}
