+++ src/main/java/duke/commands/AddEventCommand.java
@@ -1,11 +1,18 @@
-import java.io.IOException;
+package duke.commands;
 
-public class AddTodoCommand extends Command<String> {
+import java.io.IOException;
+import duke.tasks.TaskList;
+import duke.storage.Storage;
+import duke.exception.DukeException;
+import duke.tasks.Event;
+import duke.ui.Ui;
+
+public class AddEventCommand extends Command<String>{
     private TaskList list;
     private String[] echo;
     private Storage storage;
 
-    public AddTodoCommand(TaskList list, String[] echo, Storage storage) throws DukeException {
+    public AddEventCommand(TaskList list, String[] echo, Storage storage) throws DukeException {
         this.list = list;
         this.echo = echo;
         this.storage = storage;
@@ -13,7 +20,8 @@ public class AddTodoCommand extends Command<String> {
     }
 
     public void execute() throws DukeException {
-        String err = "Oh no! The description of todo cannot be empty... Try again :)\n";
+        String err = "Oh no! The description of event cannot be empty... Try again :)\n";
+        String wrongFormat = "Oh no! The format for event task is wrong... Try again :)\n";
         if (echo.length == 1) {
             throw new DukeException(err);
         }
@@ -21,7 +29,13 @@ public class AddTodoCommand extends Command<String> {
         if (description.isEmpty()) {
             throw new DukeException(err);
         }
-        Todo curr = new Todo(description);
+        String[] details = description.split(" /at ", 2);
+        if (details.length == 1) {
+            throw new DukeException(wrongFormat);
+        }
+        String info = details[0];
+        String date = details[1];
+        Event curr = new Event(info, date);
         list.addTask(curr);
         Ui.showAddResponse(curr.toString(), list.getSize());
         try {
