+++ src/main/java/duke/commands/Command.java
@@ -1,0 +1,10 @@
+package duke.commands;
+
+import duke.exception.DukeException;
+
+public abstract class Command<T> {
+    public void execute() throws DukeException, DukeException {}
+    public boolean isExit(){
+        return true;
+    }
+}
