+++ src/main/java/duke/commands/UnmarkCommand.java
@@ -1,11 +1,20 @@
+package duke.commands;
+
+import duke.commands.Command;
+import duke.tasks.TaskList;
+import duke.storage.Storage;
+import duke.exception.DukeException;
+import duke.tasks.Task;
+import duke.ui.Ui;
+
 import java.io.IOException;
 
-public class DeleteCommand extends Command<String> {
-    private TaskList list;
-    private String[] echo;
-    private Storage storage;
+public class UnmarkCommand extends Command<String> {
+    private final TaskList list;
+    private final String[] echo;
+    private final Storage storage;
 
-    public DeleteCommand(TaskList list, String[] echo, Storage storage) throws DukeException {
+    public UnmarkCommand(TaskList list, String[] echo, Storage storage) throws DukeException {
         this.list = list;
         this.echo = echo;
         this.storage = storage;
@@ -13,7 +22,7 @@ public class DeleteCommand extends Command<String> {
     }
 
     public void execute() throws DukeException {
-        String err = "Oh no! Which task do you wish to delete? Try again :)\n";
+        String err = "Oh no! Which task do you wish to unmark? Try again :)\n";
         String wrongNumber = "Oh no! This task number does not exist. Try again :)\n";
         String wrongFormat = "Oh no! Please do not spell out the number. Try again :)\n";
         int targetIndex;
@@ -34,8 +43,10 @@ public class DeleteCommand extends Command<String> {
             throw new DukeException(wrongNumber);
         } else {
             Task curr = list.getTask(targetIndex - 1);
-            list.deleteTask(targetIndex - 1);
-            Ui.showDeleteResponse(curr.toString(), size -1);
+            curr.unMark();
+            String status = curr.getStatus();
+            String description = curr.getDescription();
+            Ui.showUnmarkRes(status, description);
         }
         try {
             storage.writeToFile(list);
