+++ src/main/java/duke/tasks/Deadline.java
@@ -1,6 +1,9 @@
+package duke.tasks;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
+import duke.exception.DukeException;
 
 public class Deadline extends Task {
     protected LocalDate dueDate;
